requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==1.4.32
python-dotenv==0.18.0
psycopg2-binary==2.9.1
APScheduler==3.9.1
pandas
FlightRadarAPI
ydata_profiling
apache-airflow[celery]==2.6.0
apache-airflow-providers-common-sql>=1.3.1
psycopg2-binary>=2.8.0